pyproject.toml
[tool.poetry]
name = "donkey-see-donkey-do"
version = "0.1.0"
description = ""
authors = ["Michael Lipschultz <michael.lipschultz@gmail.com>"]
readme = "README.md"
packages = [{include = "donkey_see_donkey_do"}]

[tool.poetry.dependencies]
python = "^3.8"
pynput = "^1.7.6"
pyautogui = "^0.9.53"
pillow = "^9.4.0"
numpy = "^1.24.1"
apscheduler = "^3.10.0"
pydantic = "^1.10.4"
pin-the-tail = {git = "git@github.com:lipschultz/pin-the-tail.git", rev = "main"}

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"
pre-commit = "^2.21.0"
pytest-cov = "^4.0.0"
black = "^22.12.0"
pylint = "^2.15.10"
isort = "^5.11.4"
freezegun = "^1.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.pylint.'MESSAGES CONTROL']
disable = ["import-error", "missing-module-docstring", "missing-class-docstring", "missing-function-docstring"]

[tool.pylint.FORMAT]
max-line-length = 120

[tool.pylint.basic]
good-names = ["fp", "i", "x", "y", "dx", "dy"]
